requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
asyncer==0.0.7
bidict==0.23.1
cachetools==5.5.2
certifi==2025.1.31
chainlit==2.2.1
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cohere==5.14.0
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
fastapi==0.115.11
fastavro==1.10.0
filelock==3.17.0
filetype==1.2.0
fsspec==2025.2.0
google-auth==2.38.0
googleapis-common-protos==1.68.0
griffe==1.6.0
groq==0.18.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpath-python==1.0.6
Lazify==0.4.0
literalai==0.1.103
logfire-api==3.6.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mistralai==1.5.0
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.65.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-ai==0.0.31
pydantic-ai-slim==0.0.31
pydantic-graph==0.0.31
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
syncer==2.0.3
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
types-requests==2.32.0.20250301
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uptrace==1.29.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
